T225: Extract cmdWorkflow into workflow-cli.js
setup.js was 2041 lines — extracted the 443-line cmdWorkflow function
into workflow-cli.js. Also added workflow.js + workflow-cli.js to the
installer's core files list and --upgrade file list.

setup.js: 2041 → 1598 lines
workflow-cli.js: 452 lines (new)
